requirements.txt
elasticsearch==8.15.2
python-dotenv==1.0.1
openai==1.52.0
sentence-transformers==3.2.0